Update and rename data structures/Map and Tries/Lecture2.cpp to others/Map and Tries2.cpp
